+++ source/api_cc/include/version.h.in
@@ -16,3 +16,4 @@ const std::string global_git_date="@GIT_DATE@";
 const std::string global_git_branch="@GIT_BRANCH@";
 const std::string global_tf_include_dir="@TensorFlow_INCLUDE_DIRS@";
 const std::string global_tf_lib="@TensorFlow_LIBRARY@";
+const std::string global_model_version="@MODEL_VERSION@";
