+++ source/tests/infer/deepdipole.pbtxt
@@ -5951,6 +5951,27 @@ node {
   }
 }
 node {
+  name: "model_attr/model_version"
+  op: "Const"
+  attr {
+    key: "dtype"
+    value {
+      type: DT_STRING
+    }
+  }
+  attr {
+    key: "value"
+    value {
+      tensor {
+        dtype: DT_STRING
+        tensor_shape {
+        }
+        string_val: "1.0"
+      }
+    }
+  }
+}
+node {
   name: "model_attr/sel_type"
   op: "Const"
   attr {
