+++ source/train/run_config.ini
@@ -7,3 +7,4 @@ GIT_BRANCH = @GIT_BRANCH@
 TF_INCLUDE_DIR = @TensorFlow_INCLUDE_DIRS@
 TF_LIBS = @TensorFlow_LIBRARY@
 PRECISION = @PREC_DEF@
+MODEL_VERSION=@MODEL_VERSION@
